infra/db/importer/requirements.txt
boto3
psycopg2-binary

